+++ Clicked/Core/CommandProcessor.lua
@@ -22,6 +22,9 @@ Addon.MACRO_FRAME_HANDLER_NAME = "ClickedMacroFrameHandler"
 --- @type Button
 local macroFrameHandler
 
+--- @type boolean
+local requiresCombatProcess = false
+
 -- Local support functions
 
 --- @param frame Frame
@@ -183,6 +186,7 @@ function Addon:ProcessCommands(commands)
 
 	-- Unregister all current keybinds
 	macroFrameHandler:Hide()
+	requiresCombatProcess = false
 
 	for _, command in ipairs(commands) do
 		local attributes = {}
@@ -217,6 +221,10 @@ function Addon:ProcessCommands(commands)
 		for attribute, value in pairs(attributes) do
 			targetAttributes[attribute] = value
 		end
+
+		if (command.action == Addon.CommandType.TARGET or command.action == Addon.CommandType.MENU) and command.data ~= nil then
+			requiresCombatProcess = true
+		end
 	end
 
 	Addon:StatusOutput_UpdateMacroHandlerAttributes(newMacroFrameHandlerAttributes)
@@ -228,4 +236,11 @@ function Addon:ProcessCommands(commands)
 	Addon:StatusOutput_UpdateHovercastAttributes(newClickCastFrameAttributes)
 	Addon:UpdateClickCastHeader(newClickCastFrameKeybinds)
 	Addon:UpdateClickCastFrames(newClickCastFrameAttributes)
+end
+
+--- Get whether re-procesing of active bindings should happen when entering and leaving combat.
+---
+---@return boolean
+function Addon:IsCombatProcessRequired()
+	return requiresCombatProcess
 end